+++ src-docs/src/views/flyout/flyout_complicated.js
@@ -12,6 +12,7 @@ import {
   EuiFlyoutBody,
   EuiFlyoutFooter,
   EuiFlyoutHeader,
+  EuiPopover,
   EuiSpacer,
   EuiTab,
   EuiTabs,
@@ -27,6 +28,7 @@ export class FlyoutComplicated extends Component {
       isFlyoutVisible: false,
       isSwitchChecked: true,
       selectedTabId: '1',
+      isPopoverOpen: false,
     };
 
     this.tabs = [{
@@ -53,6 +55,14 @@ export class FlyoutComplicated extends Component {
 
   showFlyout() {
     this.setState({ isFlyoutVisible: true });
+  }
+
+  closePopover = () => {
+    this.setState({ isPopoverOpen: false });
+  }
+
+  togglePopover = () => {
+    this.setState(({ isPopoverOpen }) => ({ isPopoverOpen: !isPopoverOpen }));
   }
 
   onSelectedTabChanged = id => {
@@ -162,6 +172,13 @@ export class FlyoutComplicated extends Component {
             </EuiTabs>
           </EuiFlyoutHeader>
           <EuiFlyoutBody>
+            <EuiPopover
+              closePopover={this.closePopover}
+              button={<EuiButton onClick={this.togglePopover}>Even popovers can be included</EuiButton>}
+              isOpen={this.state.isPopoverOpen}
+            >
+              <p>This is the popover content, notice how it can overflow the flyout!</p>
+            </EuiPopover>
             {flyoutContent}
             <EuiCodeBlock language="html">
               {htmlCode}