+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Fixed disabled states of icon buttons ([#963](https://github.com/elastic/eui/pull/963))
 - Added word-break fallback for FF & IE in table cell ([#962](https://github.com/elastic/eui/pull/962))
+- Fixed `EuiPopover` to show content over modals, flyouts, etc ([#967](https://github.com/elastic/eui/pull/967))
 
 ## [`1.0.1`](https://github.com/elastic/eui/tree/v1.0.1)